Refactor the encoder implementation into CborOutputStream
Merge the encoder implementation into a single class rather than having
to instantiate 11 classes that hold the same state. This class uses an
OutputStream abstraction for outputing CBOR items to an underlying
stream.

The CborReader interface is unchanged and the CborOutputStream is not
made part of the library's public interface.
