--- src/main/java/co/nstant/in/cbor/encoder/ArrayEncoder.java
@@ -1,31 +1,0 @@
-package co.nstant.in.cbor.encoder;
-
-import java.io.OutputStream;
-import java.util.List;
-
-import co.nstant.in.cbor.CborEncoder;
-import co.nstant.in.cbor.CborException;
-import co.nstant.in.cbor.model.Array;
-import co.nstant.in.cbor.model.DataItem;
-import co.nstant.in.cbor.model.MajorType;
-
-public class ArrayEncoder extends AbstractEncoder<Array> {
-
-    public ArrayEncoder(CborEncoder encoder, OutputStream outputStream) {
-        super(encoder, outputStream);
-    }
-
-    @Override
-    public void encode(Array array) throws CborException {
-        List<DataItem> dataItems = array.getDataItems();
-        if (array.isChunked()) {
-            encodeTypeChunked(MajorType.ARRAY);
-        } else {
-            encodeTypeAndLength(MajorType.ARRAY, dataItems.size());
-        }
-        for (DataItem dataItem : dataItems) {
-            encoder.encode(dataItem);
-        }
-    }
-
-}
