--- src/main/java/co/nstant/in/cbor/encoder/TagEncoder.java
@@ -1,21 +1,0 @@
-package co.nstant.in.cbor.encoder;
-
-import java.io.OutputStream;
-
-import co.nstant.in.cbor.CborEncoder;
-import co.nstant.in.cbor.CborException;
-import co.nstant.in.cbor.model.MajorType;
-import co.nstant.in.cbor.model.Tag;
-
-public class TagEncoder extends AbstractEncoder<Tag> {
-
-    public TagEncoder(CborEncoder encoder, OutputStream outputStream) {
-        super(encoder, outputStream);
-    }
-
-    @Override
-    public void encode(Tag tag) throws CborException {
-        encodeTypeAndLength(MajorType.TAG, tag.getValue());
-    }
-
-}
