--- src/main/java/co/nstant/in/cbor/encoder/UnsignedIntegerEncoder.java
@@ -1,21 +1,0 @@
-package co.nstant.in.cbor.encoder;
-
-import java.io.OutputStream;
-
-import co.nstant.in.cbor.CborEncoder;
-import co.nstant.in.cbor.CborException;
-import co.nstant.in.cbor.model.MajorType;
-import co.nstant.in.cbor.model.UnsignedInteger;
-
-public class UnsignedIntegerEncoder extends AbstractEncoder<UnsignedInteger> {
-
-    public UnsignedIntegerEncoder(CborEncoder encoder, OutputStream outputStream) {
-        super(encoder, outputStream);
-    }
-
-    @Override
-    public void encode(UnsignedInteger dataItem) throws CborException {
-        encodeTypeAndLength(MajorType.UNSIGNED_INTEGER, dataItem.getValue());
-    }
-
-}
